refactor(metrics): split instrumentation/ into per-subsystem subdirs
ATT-276
