+++ apps/api/src/billing/sumup.service.spec.ts
@@ -11,8 +11,8 @@ import { BillingService } from './billing.service';
 import { LiveNotificationsService } from './liveNotificationsService';
 import { SumupTransactionEventType } from './dto/sumup/sumup-transaction-callback.dto';
 import { SettingsService } from '../settings/settings.service';
-import { CronTimer } from '../metrics/instrumentation/cron.helper';
-import { ExternalCallTimer } from '../metrics/instrumentation/external.helper';
+import { CronTimer } from '../metrics/instrumentation/cron/cron.helper';
+import { ExternalCallTimer } from '../metrics/instrumentation/external/external.helper';
 
 const mockSumUpGet = jest.fn();
 const mockMerchantsGet = jest.fn();
