+++ apps/api/src/mqtt/mqtt-client.service.spec.ts
@@ -8,7 +8,7 @@ import { Logger } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import { EncryptionService } from '../encryption/encryption.service';
 import { MetricsService } from '../metrics/metrics.service';
-import { ExternalCallTimer } from '../metrics/instrumentation/external.helper';
+import { ExternalCallTimer } from '../metrics/instrumentation/external/external.helper';
 
 // Interface to access private members for testing
 interface MqttClientServicePrivate {
