+++ apps/api/src/resources/flows/resource-flows-executor.service.spec.ts
@@ -18,7 +18,7 @@ import { EventEmitter2 } from '@nestjs/event-emitter';
 import { MqttMessageEvent as MqttMessageReceivedEvent } from '../../mqtt/mqtt-message.event';
 import { NoUsageSessionError } from './errors/no-usage-session.error';
 import { ResourceHealthService } from '../health/resource-health.service';
-import { CronTimer } from '../../metrics/instrumentation/cron.helper';
+import { CronTimer } from '../../metrics/instrumentation/cron/cron.helper';
 
 // Minimal edge shape for our mocks
 type Edge = { source: string; target: string; sourceHandle?: string | null };
