+++ apps/api/src/users-and-auth/auth/session.service.spec.ts
@@ -11,7 +11,7 @@ import { SessionService, SessionMetadata } from './session.service';
 import { Session, User } from '@attraccess/database-entities';
 import { TokenHashService } from '../../encryption/token-hash.service';
 import { MetricsService } from '../../metrics/metrics.service';
-import { CronTimer } from '../../metrics/instrumentation/cron.helper';
+import { CronTimer } from '../../metrics/instrumentation/cron/cron.helper';
 
 describe('SessionService', () => {
   let service: SessionService;
